+++ Wulled/Engine/src/platform/windows/WinInput.h
@@ -11,7 +11,7 @@ namespace WLD
 		virtual bool IsKeyPressedImpl(int keycode) override;
 
 		virtual bool IsMouseButtonPressedImpl(int button) override;
-		virtual std::pair<double, double> GetMousePositionImpl() override;
+		virtual std::pair<LONG, LONG> GetMousePositionImpl() override;
 		virtual double GetMouseXImpl() override;
 		virtual double GetMouseYImpl() override;
 	};
